+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 - Add toggle button for listing all served resources in server mode ([#49](https://github.com/marp-team/marp-cli/pull/49))
 - Toggle full screen by hitting <kbd>f</kbd> / <kbd>F11</kbd> in `bespoke` template ([#50](https://github.com/marp-team/marp-cli/pull/50))
 
+### Removed
+
+- Remove `util.promisify` polyfill for Node 6 ([#53](https://github.com/marp-team/marp-cli/pull/53))
+
 ## v0.0.15 - 2018-12-06
 
 ### Added